tspp: add locking to tspp_get_buffer() and tspp_release_buffer()
The tspp driver keeps a list of all the buffers and their current
state. When interrupt occurs to signal new data is available some
buffers state is updated. tspp_get_buffer and tspp_release_buffer
also change the state of buffers. This change adds locking in these
functions to protect against concurrent update which may lead to
an invalid state.

Change-Id: I6331a49bcd0a57e6b9ef8c3860bf8ca9d0e0190b
Signed-off-by: Gilad Broner <gbroner@codeaurora.org>
Signed-off-by: Monish <monishk10@yahoo.com>